Enable live MIL predictions from camera feed
